+++ service/controller/resource/namespaceauth/error.go
@@ -12,11 +12,3 @@ var invalidConfigError = &microerror.Error{
 func IsInvalidConfig(err error) bool {
 	return microerror.Cause(err) == invalidConfigError
 }
-
-var wrongTypeError = &microerror.Error{
-	Kind: "wrongTypeError",
-}
-
-func IsWrongType(err error) bool {
-	return microerror.Cause(err) == wrongTypeError
-}
